Created command_table.c/h to split parser, norm is okay
