+++ src/parser/parser.c
@@ -1,4 +1,5 @@
 #include <parser/parser.h>
+#include <parser/command_table.h>
 #include <parser/dispatcher.h>
 #include <string.h>
 #include <ctype.h>
@@ -8,91 +9,7 @@
 #include <commands/echo_handle_quotes.h>
 #include <commands/echo_utils.h>
 
-static t_bool	is_valid_last_char(const char *input, int command_len)
-{
-	const int	input_len = ft_strlen(input);
-	char		last_char;
-
-	if (input_len >= command_len)
-	{
-		last_char = input[command_len];
-		return (isspace(last_char) || last_char == '\0');
-	}
-	return (FALSE);
-}
-
-static t_bool	is_command(const char *input, const char *command)
-{
-	const int	command_len = ft_strlen(command);
-
-	return (strncmp(input, command, command_len) == 0
-		&& is_valid_last_char(input, command_len));
-}
-
-t_command_code	get_command_code(const char **input)
-{
-	static const char	*commands[LAST] = {
-											"",
-											"echo",
-											"exit",
-											"pwd",
-											"invalid"
-										};
-	t_command_code		command_code;
-
-	skip_spaces(input);
-	command_code = EMPTY_LINE;
-	while (command_code < INVALID)
-	{
-		if (is_command(*input, commands[command_code]))
-		{
-			advance_pointer(input, commands[command_code]);
-			return (command_code);
-		}
-		++command_code;
-	}
-	return (INVALID);
-}
-
-t_bool	is_between_quotes(const char *input, int reserved_char_index)
-{
-	t_quotes_index	quotes;
-	char			*new_pos;
-
-	new_pos = (char *)input;
-	while (new_pos)
-	{
-		quotes = get_quotes_indexes(new_pos);
-		if (!quotes.start && !quotes.end)
-			return (FALSE);
-		if (quotes.start < reserved_char_index
-			&& quotes.end > reserved_char_index)
-			return (TRUE);
-		new_pos = &new_pos[quotes.end + 1];
-	}
-	return (FALSE);
-}
-
-int	get_arg_len(const char *start)
-{
-	char	*pipe_position;
-	int		pipe_index;
-	int		start_index;
-
-	start_index = 0;
-	pipe_position = ft_strchr(start, PIPE);
-	while (pipe_position && start[start_index] != '\0')
-	{
-		pipe_index = pipe_position - &start[0];
-		if (!is_between_quotes(start, pipe_index))
-			return (pipe_index);
-		start_index += pipe_index + 1;
-		pipe_position = ft_strchr(&start[start_index], PIPE);
-	}
-	return (ft_strlen(start));
-}
-
-void	consume_pipe(const char **input, int index)
+static void	consume_pipe(const char **input, int index)
 {
 	if (index < 1)
 		return ;
@@ -100,7 +17,7 @@ void	consume_pipe(const char **input, int index)
 		*input += 2;
 }
 
-/* display sintax error when necessary */
+/* display syntax error when necessary */
 t_command	populate_command(const char **input_ptr)
 {
 	t_command	command;
